Added workaround for the "dlopen()-issue" of the perl plugin.
There is no way to tell lt_dlopen() to use the RTLD_GLOBAL flag which is
however required by the perl plugin (which would otherwise be unable to find
symbols defined in libperl when loading perl modules that require such
symbols).

This can be worked around by using -DLT_LAZY_OR_NOW='RTLD_LAZY|RTLD_GLOBAL'
when compiling collectd. This is rather a hack but the best solution I can
currently think of.