patches/postgresql_conf_fix.dpatch: Make sure "disk_io" doesn't return NULL.
Starting with some version between 8.3.3 and 8.3.6, pg_statio_*_tables returns
NULL instead of 0 for statistics if no instance of the appropriate relation
exists. PQgetvalue() returns an empty string in that case which would then
result in error messages when udb_result_submit() tries to convert that to a
number.

Now, the "disk_io" query uses PostgreSQL's coalesce() function to make sure 0
is returned instead of NULL.