collectd-core.collectd.init.d: Catch disabled state in start/restart.
Don't exit with an error status in that case. Amongst others, this fixes an
upgrade of collectd when the daemon is disabled.

Thanks to Florian Ernst for reporting this and Evgeni Golov for providing (an
early) patch!

To let those changes to work correctly, don't use 'set -e' and 'exit 0' (at
the end) in order to let return statuses propagate correctly.

Closes: #681216